split http_request
